Add scraping functionality for job offers and handle CSV storage
- Created TC2_formato_no_permitido.txt and TC3_archivo_vacio.xlsx for testing.
- Implemented test_scrape_home_offers in scrap_ofertas.py to scrape job offers from the home page.
- Added data extraction logic using BeautifulSoup and CSV writing for scraped offers.
- Developed test_scrape_offer in scrape_offer.py to scrape detailed offer information.
- Enhanced error handling for data extraction to ensure robustness.
