+++ backend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "sceev-backend",
-  "version": "2.3.1",
+  "version": "2.3.2",
   "private": "false",
   "description": "Backend for SCEEV project",
   "main": "index.js",
