+++ frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "sceev-frontend",
-  "version": "2.5.2",
+  "version": "2.5.3",
   "private": "false",
   "description": "Frontend for SCEEV project",
   "type": "module",
