+++ tests/data/TC2_formato_no_permitido.txt
@@ -1,0 +1,1 @@
+formato no permitido.
