+++ tests/flows/upload_test.py
@@ -1,4 +1,5 @@
 import time
+import os
 from selenium.webdriver.common.by import By
 from pages.upload_page import UploadPage
 from pages.nav_bar_components import NavbarComponent
@@ -32,7 +33,7 @@ def test_login_and_upload(driver, base_url):
 
     WebDriverWait(driver, 10).until(EC.presence_of_element_located(uploadPage.FILE_INPUT))
 
-    uploadPage.upload_file("C:\\Users\\USUARIO\\OneDrive\\Documentos\\5toSemestre\\IngenieriadeSoftwareIII\\Doc pruebas\\TC2_formato_no_permitido.txt")
+    uploadPage.upload_file(os.path.abspath(os.path.join("..", "data", "TC2_formato_no_permitido.txt")))
     time.sleep(2)
 
     uploadPage.select_municipios()
@@ -50,7 +51,7 @@ def test_login_and_upload(driver, base_url):
     except TimeoutException:
         print("Error: No se mostró el mensaje de error para archivo no permitido.")
     
-    uploadPage.upload_file("C:\\Users\\USUARIO\\OneDrive\\Documentos\\5toSemestre\\IngenieriadeSoftwareIII\\Doc pruebas\\TC3_archivo_vacio.xlsx")
+    uploadPage.upload_file(os.path.abspath(os.path.join("..", "data", "TC3_archivo_vacio.xlsx")))
     time.sleep(2)
 
     uploadPage.select_municipios()
@@ -68,7 +69,7 @@ def test_login_and_upload(driver, base_url):
     except TimeoutException:
         print("Error: No se mostró el mensaje de error para archivo vacío.")
 
-    uploadPage.upload_file("C:\\Users\\USUARIO\\OneDrive\\Documentos\\5toSemestre\\IngenieriadeSoftwareIII\\Doc pruebas\\TC1_carga_exitosa.xlsx")
+    uploadPage.upload_file(os.path.abspath(os.path.join("..", "data", "TC1_carga_exitosa.xlsx")))
     time.sleep(2)
 
     uploadPage.select_municipios()
